refactor: default to no biometric authentication
