+++ lib/screens/utilities/authenticate_screen.dart
